% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_symmetry_operations.R
\name{apply_symmetry_operations}
\alias{apply_symmetry_operations}
\title{Apply Symmetry Operations}
\usage{
apply_symmetry_operations(atomic_coordinates, symmetry_operations)
}
\arguments{
\item{atomic_coordinates}{A data.table from `extract_atomic_coordinates`.}

\item{symmetry_operations}{A data.table from `extract_symmetry_operations`.}
}
\value{
A data.table containing all unique atomic positions in the unit cell.
}
\description{
Generates all symmetry-equivalent atomic positions.
}
