% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimum_distance.R
\name{brunner}
\alias{brunner}
\title{Brunner's Largest Reciprocal Gap Method}
\usage{
brunner(distances, delta = 1e-04)
}
\arguments{
\item{distances}{data.table of interatomic distances.}

\item{delta}{Small tolerance value.}
}
\value{
A data.table of bonded pairs.
}
\description{
An alternative bonding detection method.
}
