% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_distances.R,
%   R/calculate_neighbor_counts.R
\name{calculate_angles}
\alias{calculate_angles}
\title{Calculate Bond Angles}
\usage{
calculate_angles(
  bonded_pairs,
  atomic_coordinates,
  expanded_coords,
  unit_cell_metrics
)

calculate_angles(
  bonded_pairs,
  atomic_coordinates,
  expanded_coords,
  unit_cell_metrics
)
}
\arguments{
\item{bonded_pairs}{Data.table of bonded atoms.}

\item{atomic_coordinates}{Data.table of asymmetric atom coordinates.}

\item{expanded_coords}{Data.table of supercell atom coordinates.}

\item{unit_cell_metrics}{Data.table with unit cell parameters.}
}
\value{
A data.table of all unique bond angles.

A data.table of all unique bond angles.
}
\description{
Calculates all bond angles centered on each atom.

Calculates all bond angles centered on each atom.
}
