% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_distances.R
\name{calculate_distances}
\alias{calculate_distances}
\title{Calculate Interatomic Distances}
\usage{
calculate_distances(atomic_coordinates, expanded_coords, unit_cell_metrics)
}
\arguments{
\item{atomic_coordinates}{The primary set of asymmetric atoms.}

\item{expanded_coords}{The set of atoms in the supercell.}

\item{unit_cell_metrics}{A data.table with cell parameters.}
}
\value{
A data.table of all non-zero distances.
}
\description{
Computes distances between atoms using the metric tensor.
}
