% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_symmetry_operations.R
\name{expand_transformed_coords}
\alias{expand_transformed_coords}
\title{Expand to Neighboring Unit Cells}
\usage{
expand_transformed_coords(transformed_coords)
}
\arguments{
\item{transformed_coords}{A data.table of atom positions within one unit cell.}
}
\value{
A data.table containing all atomic positions in the expanded supercell.
}
\description{
Replicates atoms into a 3x3x3 supercell grid.
}
