% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_atomic_coordinates.R,
%   R/extract_unit_cell_metrics.R
\name{extract_atomic_coordinates}
\alias{extract_atomic_coordinates}
\title{Extract Atomic Coordinates}
\usage{
extract_atomic_coordinates(cif_content)

extract_atomic_coordinates(cif_content)
}
\arguments{
\item{cif_content}{A data.table containing the lines of a CIF file.}
}
\value{
A data.table where each row corresponds to an atom.

A data.table where each row corresponds to an atom.
}
\description{
Parses atomic site information (label, coordinates, errors).

Parses atomic site information (label, coordinates, errors).
}
