% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_unit_cell_metrics.R
\name{extract_unit_cell_metrics}
\alias{extract_unit_cell_metrics}
\title{Extract Unit Cell Metrics and Errors}
\usage{
extract_unit_cell_metrics(cif_content)
}
\arguments{
\item{cif_content}{A data.table containing the lines of a CIF file.}
}
\value{
A one-row data.table with columns for each cell parameter and its error.
}
\description{
Parses unit cell parameters and their uncertainties from CIF content.
}
