% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimum_distance.R
\name{hoppe}
\alias{hoppe}
\title{Hoppe's Method of Effective Coordination Numbers}
\usage{
hoppe(distances, delta = 0.5, tolerance = 0.001)
}
\arguments{
\item{distances}{data.table of interatomic distances.}

\item{delta}{Tolerance parameter.}

\item{tolerance}{Convergence tolerance.}
}
\value{
A data.table of bonded pairs.
}
\description{
An alternative bonding detection method.
}
