% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimum_distance.R
\name{minimum_distance}
\alias{minimum_distance}
\title{Minimum Distance Method for Bonding}
\usage{
minimum_distance(distances, delta = 0.1)
}
\arguments{
\item{distances}{A data.table of interatomic distances from `calculate_distances`.}

\item{delta}{The relative tolerance parameter (default 0.1).}
}
\value{
A data.table of bonded pairs.
}
\description{
Identifies bonded atoms using the nearest neighbor distance plus a tolerance.
}
