% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propagate_distance_error.R
\name{propagate_angle_error}
\alias{propagate_angle_error}
\title{Propagate Angle Error}
\usage{
propagate_angle_error(
  bond_angles,
  atomic_coordinates,
  expanded_coords,
  unit_cell_metrics
)
}
\arguments{
\item{bond_angles}{Data.table of calculated bond angles.}

\item{atomic_coordinates}{Data.table with fractional coordinates and errors.}

\item{expanded_coords}{Data.table of supercell atom coordinates.}

\item{unit_cell_metrics}{Data.table with unit cell parameters and errors.}
}
\value{
The input `bond_angles` data.table with a new 'AngleError' column.
}
\description{
Calculates the standard uncertainty for each bond angle.
}
