% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propagate_distance_error.R
\name{propagate_distance_error}
\alias{propagate_distance_error}
\title{Propagate Distance Error}
\usage{
propagate_distance_error(bonded_pairs, atomic_coordinates, unit_cell_metrics)
}
\arguments{
\item{bonded_pairs}{Data.table of bonded atoms with their distances.}

\item{atomic_coordinates}{Data.table with fractional coordinates and errors.}

\item{unit_cell_metrics}{Data.table with unit cell parameters and errors.}
}
\value{
The input `bonded_pairs` data.table with a new 'DistanceError' column.
}
\description{
Calculates the standard uncertainty for each interatomic distance.
}
